Final fix to build link
Since this has to be a forward at the moment we must use http:// even
though it will show up as secure once the link is clicked.  Both places
the link shows up on the site now point to builds.sasview.org so in the
future if we change the link again (and we presumably will when we
retire the jenkins servers we just need to update the domain registry.