Update Gemfile to current version of github-pages
The problem is that the security risks are for versions of packages locked by the old 193 version of github-pages.  It is not clear to me whether changing this here will be sufficient or if I have to actually have to copy the repo then then run bundler on my local machine to generate the new gemfile.lock file.  This commit can always be reversed.